refactor(service): promote boards.rs to a boards/ directory module
